refactor: handler 실행 로직 변경
