+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/annotation/HandlerExecution.java
@@ -1,8 +1,9 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.annotation;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.view.JspView;
 
 import java.lang.reflect.Method;
 
@@ -17,6 +18,10 @@ public class HandlerExecution {
     }
 
     public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response) throws Exception {
-        return (ModelAndView) method.invoke(handler, request, response);
+        Object result = method.invoke(handler, request, response);
+        if (result.getClass().equals(String.class)) {
+            return new ModelAndView(new JspView((String) result));
+        }
+        return (ModelAndView) result;
     }
 }
